+++ src/components/atoms/Elements/dialogBox.tsx
@@ -7,24 +7,5 @@
 // import { Icon  } from 'office-ui-fabric-react/lib/Icon';
 
 
-export interface IMyDialogProps {
-    title: string;
-    dialogMessage: string;
-    showDialog: boolean;
-    confirmButton: string;
-    _confirmDialog: any;
-    _closeDialog: any;
-}
 
-export interface IMyBigDialogProps {
-  title: string;
-  dialogMessage?: string;
-  dialogElement?: any;
-  showDialog: boolean;
-  confirmButton: string;
-  _confirmDialog: any;
-  _closeDialog: any;
-  minWidth?: number;
-  maxWidth?: number;
-}