+++ src/logic/indexes/DOMSearch.ts
@@ -1,5 +1,5 @@
 
-import type { findParentElementLikeThis, findParentElementPropLikeThis } from "../DOM/Search/domSearch";
+import { findParentElementLikeThis, findParentElementPropLikeThis } from "../DOM/Search/domSearch";
 
 export { findParentElementLikeThis, findParentElementPropLikeThis, };