+++ CHANGELOG.md
@@ -12,6 +12,14 @@ npm install @mikezimm/fps-library-v2@1.0.38
 
 NOTE:  TURN BACK ON STRICT NULL CHECKS - turned off for PropPaneCols
 
+## Part 2
+- move replaceHandleBars from src/logic/Strings/handleBars.ts to /handleBarsContext.ts to
+
+## 1.0.39 - 2022-Dec-16:  Drilldown Update
+- npm install @mikezimm/fps-pnp2@1.0.13 >> Has Drilldown fetch and update core function, wrapper function to be here.
+- Create ReactList View from Drilldown: src/components/molecules/ReactListV1
+- rename handleBars functions
+
 ## 1.0.39 - 2022-Dec-16:  Drilldown Update
 - Migrate Drilldown PropPaneHelp pages - src/common/PropPaneHelp/pages/listview
 - Restructured PropPaneHelp folder into subfolders